requirements.txt
Flask>=2.0.0
duckdb>=0.6.0
feedparser
requests
beautifulsoup4
sumy
langdetect
sentence-transformers>=2.2.2
faiss-cpu>=1.7.1
ollama>=0.0.1
kafka-python
rank-bm25>=0.2.2
hf_xet
aiohttp
pandas
numpy
prometheus-client
google-cloud-bigquery
rake_nltk
pytest

